refactor: Прочие мелкие правки
